+++ source/libs/monitorfw/src/taos_metric_sample.c
@@ -112,7 +112,7 @@ int taos_metric_sample_set(taos_metric_sample_t *self, double r_value) {
   return 0;
 }
 
-int taos_metric_sample_exchange(taos_metric_sample_t *self, double r_value, double* old_value) {
+int taos_metric_sample_exchange(taos_metric_sample_t *self, int64_t r_value, int64_t* old_value) {
   if (self->type != TAOS_GAUGE && self->type != TAOS_COUNTER) {
     TAOS_LOG(TAOS_METRIC_INCORRECT_TYPE);
     return 1;
